Add pseudo-logout feature
It's a fake because we aren't really doing anything. The front end is expected to handle logout. That said, if the front end has the token, it is considered logged in as the backend is concerned. Therefore, logging out could be implemented by deleting the jwt token or something close'