Update GET /cart
It was initially GET /cart, now it can be accessed through GET /users/:id/cart.
This is to ensure the user id has been provided so it can be used to determine
whose cart we want to get. I have also changed the logic inside it to ensure
we are actually getting that persons's cart and not another's which was a
possibility with the previous approach