Add validators to shopping cart model
One validator ensures that the quantity never goes below 1.
There should always be more or one of a product in the cart

The other validator ensures that there is never more than one
combination of product_id and order_id in the shopping_cart.
For instance, if one shopping cart is of the form
{product_id: 1, order_id: 1, quantity: 1}, there should never
be another in the form {product_id: 1, quantity_id: 1, quantity: 3}
regardless of the quantity. This forces changing quantity to involve
updating the quantity and not posting a new shopping_cart